pallet_commitment::helpers

Function finalize_resolve_commit

fn finalize_resolve_commit<T: Config<I>, I: 'static>(
    who: &Proprietor<T>,
    reason: &<Pallet<T, I> as Commitment<Proprietor<T>>>::Reason,
    imbalance: AssetDelta<T, I>,
) -> Result<<Pallet<T, I> as InspectAsset<Proprietor<T>>>::Asset, DispatchError>

Low-level helper that centralizes the finalization logic for resolving commits.

This function is used across direct, index, and pool digests and serves as the final step of the resolving an existing commit’s operation.

Highly unchecked and assumes all invariants are already validated. Therefore, it is kept private and scoped to this module.