pallet_commitment::pallet

Enum Event

pub enum Event<T: Config<I>, I: 'static = ()> {

    CommitPlaced {
        who: Proprietor<T>,
        reason: CommitReason<T, I>,
        digest: Digest<T>,
        value: AssetOf<T, I>,
        variant: T::Position,
    },
    CommitRaised {
        who: Proprietor<T>,
        reason: CommitReason<T, I>,
        digest: Digest<T>,
        value: AssetOf<T, I>,
    },
    CommitResolved {
        who: Proprietor<T>,
        reason: CommitReason<T, I>,
        digest: Digest<T>,
        value: AssetOf<T, I>,
    },
    CommitValue {
        digest: Digest<T>,
        reason: CommitReason<T, I>,
        value: AssetOf<T, I>,
    },
    DigestInfo {
        digest: Digest<T>,
        reason: CommitReason<T, I>,
        value: AssetOf<T, I>,
        variant: T::Position,
    },
    DigestReaped {
        digest: Digest<T>,
        reason: CommitReason<T, I>,
        dust: AssetOf<T, I>,
    },
    IndexInitialized {
        index_of: IndexDigest<T>,
        reason: CommitReason<T, I>,
    },
    IndexValue {
        index_of: IndexDigest<T>,
        reason: CommitReason<T, I>,
        value: AssetOf<T, I>,
    },
    IndexEntryValue {
        index_of: IndexDigest<T>,
        reason: CommitReason<T, I>,
        entry_of: Digest<T>,
        value: AssetOf<T, I>,
    },
    IndexEntriesValue {
        index_of: IndexDigest<T>,
        reason: CommitReason<T, I>,
        entries: Vec<(EntryDigest<T>, AssetOf<T, I>)>,
    },
    IndexReaped {
        index_of: IndexDigest<T>,
        reason: CommitReason<T, I>,
    },
    PoolManager {
        pool_of: PoolDigest<T>,
        reason: CommitReason<T, I>,
        manager: Proprietor<T>,
    },
    PoolInitialized {
        pool_of: PoolDigest<T>,
        reason: CommitReason<T, I>,
        commission: T::Commission,
    },
    PoolSlot {
        pool_of: PoolDigest<T>,
        reason: CommitReason<T, I>,
        slot_of: SlotDigest<T>,
        variant: T::Position,
        shares: T::Shares,
    },
    PoolValue {
        pool_of: PoolDigest<T>,
        reason: CommitReason<T, I>,
        value: AssetOf<T, I>,
    },
    PoolSlotValue {
        pool_of: PoolDigest<T>,
        reason: CommitReason<T, I>,
        slot_of: SlotDigest<T>,
        value: AssetOf<T, I>,
    },
    PoolSlotsValue {
        pool_of: PoolDigest<T>,
        reason: CommitReason<T, I>,
        slots: Vec<(SlotDigest<T>, AssetOf<T, I>)>,
    },
    PoolCommission {
        pool_of: PoolDigest<T>,
        reason: CommitReason<T, I>,
        commission: T::Commission,
    },
    ReserveDeposited {
        amount: AssetOf<T, I>,
        total_on_hold: AssetOf<T, I>,
    },
    ReserveWithdrawn {
        amount: AssetOf<T, I>,
        total_on_hold: AssetOf<T, I>,
    },
    PoolReaped {
        pool_of: PoolDigest<T>,
        reason: CommitReason<T, I>,
    },
    PoolSlotRemoved {
        pool_of: PoolDigest<T>,
        reason: CommitReason<T, I>,
        slot_of: SlotDigest<T>,
        variant: T::Position,
    },
    DigestModel {
        digest: DigestVariant<T, I>,
    },
    AssetIssuable {
        asset: AssetOf<T, I>,
    },
    AssetReapable {
        asset: AssetOf<T, I>,
    },
    ReasonValuation {
        reason: CommitReason<T, I>,
        value: AssetOf<T, I>,
    },
    // some variants omitted

}

The Event enum of this pallet

Variants

CommitPlaced

Emitted when a proprietor places a new commit on a digest with a specific variant.

Fields

who: Proprietor<T>

reason: CommitReason<T, I>

digest: Digest<T>

value: AssetOf<T, I>

variant: T::Position

CommitRaised

Emitted when an existing commit for a digest is increased or raised.

Fields

who: Proprietor<T>

reason: CommitReason<T, I>

digest: Digest<T>

value: AssetOf<T, I>

CommitResolved

Emitted when a commit is resolved (finalized/withdrawn) for a digest.

Fields

who: Proprietor<T>

reason: CommitReason<T, I>

digest: Digest<T>

value: AssetOf<T, I>

CommitValue

Emitted when querying the current committed value for a specific digest and reason.

Fields

digest: Digest<T>

reason: CommitReason<T, I>

value: AssetOf<T, I>

DigestInfo

Emitted when the effective value of the digest variant is updated.

Fields

digest: Digest<T>

reason: CommitReason<T, I>

value: AssetOf<T, I>

variant: T::Position

DigestReaped

Emitted when a direct-digest is reaped (removed) after all commitments are cleared from it.

dust represents unclaimable dead asset value.

Fields

digest: Digest<T>

reason: CommitReason<T, I>

dust: AssetOf<T, I>

IndexInitialized

Emitted when a new index of variants is initialized. Contains all entries (digests, sharesa and variant) of the index.

Fields

index_of: IndexDigest<T>

reason: CommitReason<T, I>

IndexValue

Emitted when querying the total value of an index for a specific proprietor.

Fields

index_of: IndexDigest<T>

reason: CommitReason<T, I>

value: AssetOf<T, I>

IndexEntryValue

Emitted when querying the value of a specific entry within an index.

Fields

index_of: IndexDigest<T>

reason: CommitReason<T, I>

entry_of: Digest<T>

value: AssetOf<T, I>

IndexEntriesValue

Emitted when querying the values of all entries within an index.

Fields

index_of: IndexDigest<T>

reason: CommitReason<T, I>

entries: Vec<(EntryDigest<T>, AssetOf<T, I>)>

IndexReaped

Emitted when a index is reaped (removed) after all entry commitments are cleared from it.

Fields

index_of: IndexDigest<T>

reason: CommitReason<T, I>

PoolManager

Emitted when a pool’s manager is set or updated. The manager is responsible for managing slots and internal pool operations.

Fields

pool_of: PoolDigest<T>

reason: CommitReason<T, I>

manager: Proprietor<T>

PoolInitialized

Emitted when a new pool is initialized from an index. Includes the commission rate and initial slots with their associated shares and variants.

Fields

pool_of: PoolDigest<T>

reason: CommitReason<T, I>

commission: T::Commission

PoolSlot

Emitted when a slot within a pool has its variant updated or a new slot is added.

Fields

pool_of: PoolDigest<T>

reason: CommitReason<T, I>

slot_of: SlotDigest<T>

variant: T::Position

shares: T::Shares

PoolValue

Emitted when querying the total value of a pool for a specific proprietor.

Fields

pool_of: PoolDigest<T>

reason: CommitReason<T, I>

value: AssetOf<T, I>

PoolSlotValue

Emitted when querying the value of a specific slot within a pool.

Fields

pool_of: PoolDigest<T>

reason: CommitReason<T, I>

slot_of: SlotDigest<T>

value: AssetOf<T, I>

PoolSlotsValue

Emitted when querying the values of all slots within a pool.

Fields

pool_of: PoolDigest<T>

reason: CommitReason<T, I>

slots: Vec<(SlotDigest<T>, AssetOf<T, I>)>

PoolCommission

Emitted when querying or updating the commission rate of a pool.

Fields

pool_of: PoolDigest<T>

reason: CommitReason<T, I>

commission: T::Commission

ReserveDeposited

Emitted when funds are deposited into reserve (held balance) under the prepare-for-commit reason.

Fields

amount: AssetOf<T, I>

total_on_hold: AssetOf<T, I>

ReserveWithdrawn

Emitted when reserved funds are withdrawn back to free balance.

Fields

amount: AssetOf<T, I>

total_on_hold: AssetOf<T, I>

PoolReaped

Emitted when a pool is reaped (removed) after all slot commitments are cleared from it.

Fields

pool_of: PoolDigest<T>

reason: CommitReason<T, I>

PoolSlotRemoved

Emitted when a pool slot is removed due to its shares being zero.

Fields

pool_of: PoolDigest<T>

reason: CommitReason<T, I>

slot_of: SlotDigest<T>

variant: T::Position

DigestModel

Emitted when determining the digest model (Direct, Index, or Pool) for a given digest.

Fields

digest: DigestVariant<T, I>

AssetIssuable

Emitted when the total assets pending issuance are queried.

Fields

asset: AssetOf<T, I>

AssetReapable

Emitted when the total assets pending reaping are queried.

Fields

asset: AssetOf<T, I>

ReasonValuation

Emitted when the total committed value for a reason is queried.

Fields

reason: CommitReason<T, I>

value: AssetOf<T, I>