pallet_commitment::pallet

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AssetToIssue

Tracks the total amount of assets that are scheduled to be issued/minted.

AssetToReap

Tracks the total amount of assets that are scheduled to be reaped/burned.

BalanceSnapShots

Snapshot storage for LazyBalance state.

CommitMap

Maps each Proprietor and the committed reason (CommitReason) to their commitment information.

DigestMap

Maps each CommitReason and its associated Digest to the digest’s information.

EntryMap

Stores commit information for individual entries within an index, scoped by both a CommitReason and the corresponding Proprietor.

IndexMap

Stores index information for a given CommitReason additionally keyed by IndexDigest.

PoolManager

Tracks the manager of each pool for a given reason.

PoolMap

Stores pool-level information associated with a specific CommitReason.

ReasonValue

Tracks the total committed asset value for a specific CommitReason.