pallet_commitment::pallet

Type Alias AssetToReap

pub type AssetToReap<T: Config<I>, I: 'static = ()> = StorageValue<_GeneratedPrefixForStorageAssetToReap<T, I>, AssetOf<T, I>, ValueQuery>;

Tracks the total amount of assets that are scheduled to be reaped/burned.

This value is updated whenever commitments decrease a digest’s balance (penalty/deflation).

Similar to AssetToIssue, these operations are lazy and may not immediately affect the underlying
base asset pallet. AssetToReap provides an accounting view of the total assets that will
soon be removed eventually from circulation, ensuring proper bookkeeping and equilibrium.

Storage type is [StorageValue] with value type AssetOf < T, I >.

Aliased Type

pub struct AssetToReap<T: Config<I>, I: 'static = ()>(/* private fields */);