pallet_commitment::pallet

Type Alias DigestMap

pub type DigestMap<T: Config<I>, I: 'static = ()> = StorageNMap<_GeneratedPrefixForStorageDigestMap<T, I>, (NMapKey<Blake2_128Concat, CommitReason<T, I>>, NMapKey<Blake2_128Concat, Digest<T>>), DigestInfo<T, I>, OptionQuery>;

Maps each CommitReason and its associated Digest to the digest’s information.

Key Structure:

(CommitReason, Digest) -> DigestInfo

This storage ensures that every digest is always tied to a reason. A digest cannot exist independently without a reason, enforcing a strict relationship between commitments and their context.

Storage type is [StorageNMap] with keys type (CommitReason < T, I >, Digest < T >) and value type DigestInfo < T, I >.

Aliased Type

pub struct DigestMap<T: Config<I>, I: 'static = ()>(/* private fields */);