pallet_commitment::pallet

Type Alias PoolManager

pub type PoolManager<T: Config<I>, I: 'static = ()> = StorageNMap<_GeneratedPrefixForStoragePoolManager<T, I>, (NMapKey<Blake2_128Concat, CommitReason<T, I>>, NMapKey<Blake2_128Concat, PoolDigest<T>>), Proprietor<T>, OptionQuery>;

Tracks the manager of each pool for a given reason.

Each pool has a designated manager (proprietor) responsible for slot management. Pools are mutable, but their manager can be transfered/updated.

Keys:

CommitReason - the reason/context under which the pool exists.

PoolDigest - the unique identifier of the pool.

Value: Proprietor - account managing the pool.

Query behavior: OptionQuery returns None if the pool has no assigned manager, effectively stale.

Storage type is [StorageNMap] with keys type (CommitReason < T, I >, PoolDigest < T >) and value type Proprietor < T >.

Aliased Type

pub struct PoolManager<T: Config<I>, I: 'static = ()>(/* private fields */);